Added all classes and deleted all unneeded classes except: Empty class is not needed (the code within I do not know, but the class itself is not needed). Motion to delete once everyone is ready.
Added all instance variables and blank methods for the controller, Game, and Team classes. I left off on inputting the variables and methods for the Board class, as I will go study for our test tonight, but since defining methods and constructors for inherited classes is a bit more involved (and lacking in experience), I didn't have time to finish declaring all of the objects that inherit from class Piece in the Board class, nor did I get to inputting all of Piece class's and its subclasses' members, for this reason. But this will be all we have to do to have all of the instance variables and methods input into the program. Then we will just need to code the algorithms.