Fixed FXML file error: for some reason the C8 click event was set to call C1's click event.
I added an extra feature: sounds for when the pieces move. There are four sounds: when a piece simply moves, when a piece captures another piece, when a piece is moved and the move results in placing the enemy king in check, and when a king castles. These new methods are added in the TeamLewisChess main class; they are static methods and can be called in the Board class when a specific kind of move is made. Two of these methods are also called in the NewGameMenuItemClicked method.

Finally, I set the GUI window's resizable property to false so that the user can't resize the window and see that the chess board is just an image anchored to the upper left (which could be fixed to be centered or something if we had it contained in an anchor pane or something, but this will do for now.).