Committing with the three new files for the Team Lewis Chess Welcome Screen that Adam and I started: its .java file, .FXML file, and its controller class. We just couldn't get it to compile for some reason, though I did compile the new window for the pawn recently, so I know it must be something small.
The Promotion controller, and with that all of the pawn promotion functionality, is working. The two previous tasks are complete: updated pawn's promotion variables and the promotion screen's event handler methods. The third step of the updateBoard method is complete. Will finish up the first two steps and then input the piece move, capture, check, and castle sounds in their proper spots.