I completed methods updateBoard, isACastlingMove, isAnEnPassantMove, and updateGameStatus. The isBeingPromoted method was not needed and so erased. I had to add a few setter functions for the Game and controller class for the updateBoard method.
I also found a duct tape fix for now for the sound files when building the application: simply make sure the .jar file is in the same directory as the sound files. My project wouldn't allow me to add a resources file from the project properties menu to place the .wav files in so that I could use a specific method for retrieving those resources--I thought it could be the Netbeans build, but I don't know--so that'll do for now, so that each of us can keep our chess program forevers.

All the sound methods are all called in the Board class now, just to make it easy to know where they all are: in the beginning of the resetBoardForNewGame method and at the end of the updateBoard method.

In summary, the players (using two mice or sharing the same mouse) may play through a whole game of chess, so long as they do not make any illegal moves, which also means they cannot accidentally move in check. So, the methods pending to completion of the program are the isValidMove methods and the King's isInCheck method. The King object's isInCheck method is used for updating the game status in the updateGameStatus method and in playing the check sound when a king is put in check. Castling can occur and en passant can occur. Again, their moves just must now be validated in the isValidMove methods. When one player places another in checkmate, once these methods are complete, the game will determine the result, display a message, and freeze the game so that no more pieces may be updated with new moves, since the game is over; this is true for stalemate and a second kind of draw: draw by the Fifty Move Rule. The three other types of draws, as pre-planned and further described in the Tasks list and algorithm, are "extra features" for now. They could be programmed given another day or a little less.