Implemented capturing en passant. Fixed bug in Board class that prevented the squareMovedTwoSpacesLastMove boolean from resetting to false after a pawn had moved twice and it was tripped to true. Now the Pawn piece is completely done. For the sake of our project, the project is as complete as is our goal. The only malfunctions or incomplete features that make our chess program not wholly simulating all the rules of a chess game are these:
When the updateGameStatus method is not commented out, it does not check for draw by repetition, draw by perpetual check, nor draw by insufficient material, as these solutions have not been thought out yet. And also, while the updateGameStatus method does successfully check for checkmate, there is a bug when checking for stalemate where the king "walks into stalemate," and the method updates the Game Status to stalemate when the other player still has a move to make. And finally, the reason that updateGameStatus is commented out is because, by the design of iterating through the board array to find all the pieces and then check for if there is a valid move, this slows down the program considerably and is just not a feasible implementation; so this method is commented out so that the program does not run slow.

In conclusion, these are the deficiencies our project ended with, but other than these, our project simulates the rules of a real chess game and all of the pieces abide by these rules; when checkmate is reached, although a message does not display because the updateGameStatus method is commented out, checkmate is found to be evident as the player who is in checkmate will not be able to move another piece, and the king is never able to be captured because of the complete implementation of all of the pieces' moveDoesNotPlaceKingInCheck methods, and so no arrayOutOfBoundsException is ever thrown; our program is robust in that no situation can come about on the board where an arrayOutOfBoundsException error or some other error occurs.