Jolly!!! First, I was playing through a game to check for bugs and found a bug with the queen and rook. So I replaced Ms. Peggy's code with Leana's, since I understood Leana's code but I don't understand how the array lists work. The first jolly news is that I found the bug with Leana's queen and rook code! So I fixed that. After commenting out the debug comments that returned when moving a piece if each sub-method returned true or not, for each piece, I decided to try to reinstate the updateGameStatus method, and guess what? The second Jolly news: it doesn't slow down the game!!! The only slow point is when it is indeed checkmate or stalemate, and it is verifying it. So, our algorithm would not be optimal for a speed chess program, but other than that, it is not annoyingly slow every turn! So then, I went and tried to fix the bug when checking for stalemate. First, I found a bug in that I could take a rook, then leave that square, and the king could castle anyway, so I fixed that. So then, I fixed the stalemate bug as well! That's the third jollyness!!!
The Jollyness Summary: We have a complete chess program except for those three "extra features" that could be programmed given just a little more time, where the game checks for the Game Statuses of draw by repetition, draw by perpetual check, and draw by insufficient material! But that's it! The game, to my knowledge and through what I believe has been sufficient testing for potential problems, nearly completely simulates a legal chess game (again, except for those last Game Statuses that it could check for), and it is not slow except when verifying checkmate or stalemate!!!